feat/structure: separated components from containers
