+++ src/actions/index.js
@@ -1,16 +1,16 @@
-import * as Types from './types';
+import * as constants from '../constants';
 
-export const changeEmail = (email) => ({
-    type: Types.CHANGE_EMAIL,
-    payload: {email},
+export const changeEmail = email => ({
+  type: constants.CHANGE_EMAIL,
+  payload: { email },
 });
 
-export const changePassword = (password) => ({
-    type: Types.CHANGE_PASSWORD,
-    payload: {password},
+export const changePassword = password => ({
+  type: constants.CHANGE_PASSWORD,
+  payload: { password },
 });
 
 export const submit = () => ({
-    type: Types.SUBMIT,
-    payload: {},
+  type: constants.SUBMIT,
+  payload: {},
 });
