--- src/actions/types.js
@@ -1,3 +1,0 @@
-export const CHANGE_EMAIL = 'CHANGE_EMAIL';
-export const CHANGE_PASSWORD = 'CHANGE_PASSWORD';
-export const SUBMIT = 'SUBMIT';
