+++ src/index.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { createStore } from 'redux';
+import { Provider } from 'react-redux';
 import reducers from './reducers';
-import {createStore} from "redux";
-import {Provider} from 'react-redux';
 import App from './components/App';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
-    <Provider store={createStore(reducers)}>
-        <App />
-    </Provider>,
-    document.getElementById('root'),
+  <Provider store={createStore(reducers)}>
+    <App />
+  </Provider>,
+  document.getElementById('root'),
 );
 
 // If you want your app to work offline and load faster, you can change
