+++ src/selectors/index.js
@@ -1,23 +1,20 @@
-import {createSelector} from "reselect";
+import { createSelector } from 'reselect';
+import { EMAIL_REGEXP } from '../constants';
 
-const emailRegexp =
-    // eslint-disable-next-line
-    /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-
-const emailSelector = state => state.get("email");
-const passwordSelector = state => state.get("password");
+const emailSelector = state => state.get('email');
+const passwordSelector = state => state.get('password');
 
 export const emailIsValidSelector = createSelector(
-    emailSelector,
-    email => emailRegexp.test(email),
+  emailSelector,
+  email => EMAIL_REGEXP.test(email),
 );
 
 export const passwordIsValidSelector = createSelector(
-    passwordSelector,
-    password => password.length > 4,
+  passwordSelector,
+  password => password.length > 4,
 );
 
 export const isSubmittable = createSelector(
-    [emailIsValidSelector, passwordIsValidSelector],
-    (emailIsValid, passwordIsValid) => emailIsValid && passwordIsValid,
+  [emailIsValidSelector, passwordIsValidSelector],
+  (emailIsValid, passwordIsValid) => emailIsValid && passwordIsValid,
 );
