+++ src/styles/Email.css
@@ -5,10 +5,6 @@
     align-items: stretch;
 }
 
-.EmailContainer label {
-
-}
-
 .EmailContainer .Email {
     color: #1189FC;
     border: 1px solid #99CDFF;
