+++ src/styles/Submit.css
@@ -1,7 +1,3 @@
-.SubmitContainer {
-
-}
-
 .SubmitContainer .SubmitHelperText {
     text-align: center;
 }
